Put property 'Exp' component into own file
